+++ OpenMarket/OpenMarket/Models/Network/Protocol/MultiPartForm.swift
@@ -1,0 +1,11 @@
+//
+//  MultiPartForm.swift
+//  OpenMarket
+//
+//  Created by Fezravien on 2021/09/12.
+//
+
+import Foundation
+
+protocol MultiPartForm { }
+
